Created the Movie model
