+++ movieDeets/movieDeets/MODELS/SearchResult.swift
@@ -21,19 +21,3 @@ struct SearchResult: Codable {
         case error = "Error"
     }
 }
-
-struct SearchResultItem: Codable {
-    let title: String?
-    let year: String?
-    let imdbID: String?
-    let type: String?
-    let poster: String?
-    
-    enum CodingKeys: String, CodingKey {
-        case title = "Title"
-        case year = "Year"
-        case imdbID
-        case type = "Type"
-        case poster = "Poster"
-    }
-}
